+++ packages/keycloak/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.0](https://github.com/skore-io/nestjs-extensions/compare/v0.15.1...v0.16.0) (2020-12-02)
+
+
+### Features 🚀
+
+* **keycloak:** update resource ([#27](https://github.com/skore-io/nestjs-extensions/issues/27)) ([8c0efc2](https://github.com/skore-io/nestjs-extensions/commit/8c0efc25b02f993901901eed9337b89ba1fbe055))
+
+
+
+### [0.15.1](https://github.com/skore-io/nestjs-extensions/compare/v0.15.0...v0.15.1) (2020-12-02)
+
+
+### Bug Fixes 🐛
+
+* **@skore-io/keycloak:** exports DeleteResourceService ([#28](https://github.com/skore-io/nestjs-extensions/issues/28)) ([a211a43](https://github.com/skore-io/nestjs-extensions/commit/a211a438490e9bf284f3a91c4d0b571519b2be8d))
+
+
+
 ## [0.15.0](https://github.com/skore-io/nestjs-extensions/compare/v0.14.0...v0.15.0) (2020-12-01)